sloppy commit - add scaled dot
