+++ src/layers/__init__.py
@@ -1,6 +1,6 @@
-from add_norm import AddNorm
+from scaled_dot import ScaledDotProduct
 from feed_forward import FeedForward
 from multihead import MultiHeadAttention
 from input_encoding import InputEncoding
 from output_embedding import OutputEmbedding
-from attention import Attention
+from src.layers.scaled_dot import ScaledDotProduct
